sql scripts updated. write-only table creation script created. create_tables.sql valence type fixed (int to float). inset_tweets.sql bug fix.

+++ sql/create_tables.sql
@@ -19,7 +19,7 @@ CREATE TABLE DATA_survey_results (
    user_id INT NOT NULL,
    tweet_id INT NOT NULL,
    twitter_id VARCHAR(32) NOT NULL,
-   valence INT NOT NULL,
+   valence FLOAT NOT NULL,
    class_id INT NOT NULL,
    PRIMARY KEY (user_id, tweet_id),
    FOREIGN KEY (user_id) REFERENCES META_survey_users(id),
